// TPT/manage-tradingpt/app/api/users.ts
export interface UidInfo {
  exchangeName: string;
  uid: string;
}

// 회원 검색 결과 인터페이스
export interface SearchedUser {
  userId: number;
  name: string;
  phone: string;
  uid: UidInfo | null;
  status: string;
  requestedAt: string;
  trainerName: string | null;
  token: number;
}